requirements.txt
asgiref==3.3.4
astroid==2.5.3
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
dj-database-url==0.5.0
Django==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-postgresql==0.0.3
django-render-partial==0.4
djangorestframework==3.12.4
dnspython==2.1.0
docopt==0.6.2
dodgy==0.2.1
extras==1.0.0
fixtures==3.0.0
flake8==3.9.0
flake8-polyfill==1.0.2
gunicorn==20.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
linecache2==1.0.0
Markdown==3.3.4
mccabe==0.6.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
pbr==5.5.1
pep8-naming==0.11.1
pgbouncer==0.0.9
pipreqs==0.4.10
protobuf==3.15.8
psycopg2==2.8.6
pycodestyle==2.7.0
pydocstyle==6.0.0
pyflakes==2.3.1
pylint==2.7.4
pylint-celery==0.3
pylint-django==2.4.3
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dotenv==0.17.0
python-mimeparse==1.6.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requirements-detector==0.7
setoptconf==0.2.0
six==1.15.0
snowballstemmer==2.1.0
sqlparse==0.4.1
testtools==2.4.0
toml==0.10.2
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
yarg==0.1.9
